+++ crates/mongodb-connector/src/mongo_connector.rs
@@ -38,10 +38,11 @@ impl ConnectorSetup for MongoConnector {
             .map_err(|err| {
                 ErrorResponse::new(
                     StatusCode::INTERNAL_SERVER_ERROR,
-                    err.to_string(),
+                    format!("{err:#}"), // alternate selector (:#) includes root cause in string
                     json!({}),
                 )
             })?;
+        tracing::debug!(?configuration);
         Ok(MongoConfiguration(configuration))
     }