+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog documents the changes between release versions.
 ### Fixed
 
 - Fixes for filtering by complex predicate that references variables, or field names that require escaping ([#111](https://github.com/hasura/ndc-mongodb/pull/111))
+- Escape names if necessary instead of failing when joining relationship on field names with special characters ([#113](https://github.com/hasura/ndc-mongodb/pull/113))
 
 ## [1.3.0] - 2024-10-01